requirements.txt
# Core dependencies
python-dotenv>=1.0.1
langchain>=0.3.19
langchain-community>=0.3.18
langchain-core>=0.3.40
langchain-openai>=0.3.7
langchain-text-splitters>=0.3.6
langchain-milvus>=0.1.1
langchain-xai>=0.1.0

# OpenAI
openai>=1.64.0

# UI
streamlit>=1.42.2

# Document processing
pypdf>=5.3.0
docx2txt>=0.8
python-pptx>=0.6.21
unstructured>=0.12.0

# Vector store
pymilvus>=2.3.0

# Utils
tiktoken>=0.9.0  # Needed for token counting with OpenAI
aiohttp>=3.8.3  # Required by langchain

